Add blog page and update favicon path
Created a new blog.html page with a list of blog resources. Moved favicon.ico from assets/ to the root directory and updated its reference in index.html. Modified index.html to display a list of external links with icons and moved the previous blog links to the new blog.html.